refactor: extract DAM log rendering helpers
Move dam logs filtering, summary shaping, and text rendering out of main.rs into a focused module while preserving the existing CLI behavior.
