+++ docs/dam.md
@@ -140,7 +140,7 @@ The previous one-shot `npx @rpblc/dam claude` and `npx @rpblc/dam codex --api` t
 - `dam doctor --json`, `dam setup status --json`, `dam setup plan --json`, `dam setup next-action --json`, and `dam setup resume --json` are the headless install/resume contract. They do not mutate system state and include command tokens, confirmation flags, and `changes_system` for the next action. `doctor`, `status`, `plan`, `next-action`, and `resume` accept the same setup `--network-mode` and `--trust-mode` flags so agents can inspect explicit-proxy, system-proxy, or Network Extension release paths consistently. `status` and `plan` return the same full checklist. Each setup planning subcommand prints command-specific `--help` output instead of a combined synopsis so humans and agents can copy the exact recovery command they asked about. Each setup step includes stable `kind`, `status`, and `detail` fields for machine callers; English `message` text remains diagnostic/support copy only.
 - `dam setup rescue` is the local recovery contract. It previews by default; `--yes` stops the DAM daemon if needed and removes DAM-managed macOS system proxy and Network Extension routing state so normal networking can resume. It accepts `--state-dir` for support and test sessions. It does not remove local CA trust or delete the vault. `dam setup repair` wraps the same rescue preview/apply result together with a fresh setup plan, and `dam setup export-diagnostics` emits an offline doctor/setup/rescue-preview bundle for support and autonomous installers. The same payload family is exposed through `/api/v1/setup/rescue`, `/api/v1/setup/repair`, `/api/v1/setup/diagnostics`, `dam_setup_rescue`, `dam_setup_repair`, and `dam_setup_export_diagnostics` for local agents; mutating API/MCP rescue or repair calls require the `remove_dam_network_setup` confirmation string.
 - `dam connect --json` and `dam disconnect --json` return stable machine-readable lifecycle results for agent and script callers.
-- `dam logs` reads the local SQLite log and renders concise operation summaries by default. `--operation <id>` shows one operation's event timeline, and `--json` keeps the same data machine-readable for local debugging. Activity values live in the log store for the web Activity feed, while the CLI summary avoids raw request/response bodies.
+- `dam logs` reads the local SQLite log and renders concise operation summaries by default. `--operation <id>` shows one operation's event timeline, and `--json` keeps the same data machine-readable for local debugging. Activity values live in the log store for the web Activity feed, while the CLI summary avoids raw request/response bodies. CLI log filtering, summary shaping, and text rendering live in `crates/dam/src/log_rendering.rs`; `main.rs` only wires the command to the log store and serializer.
 - `dam disconnect` pauses protection without stopping the daemon. `dam connect` resumes a paused daemon using its existing routing/trust setup. If the connected daemon was launched by a missing or different `dam` executable path/fingerprint, Connect restarts it from the current executable while preserving that setup, so source builds and app updates do not keep running stale proxy code. Use `dam disconnect --stop` before intentionally changing setup.
 - `dam profile set <id>` persists the legacy active local harness profile. The tray/web Settings flow persists enabled app profiles; when no state exists, DAM enables every loaded profile.
 - `dam connect --network-mode system_proxy` refuses to start until DAM sees macOS PAC routing installed. Run `dam network install-system-proxy --yes` first after reviewing the preview.
